Move curses sim ui to simulate command file
